TOOLS: Move navigator tools over to mil_tools
